Expected<T&> now stores T& internally
I don't know if this is a good idea. But now the structure technically matches the user's intuitions/intent. If the user wants to return a reference to an existing structure, now the can.  It will help not copy large structures, which I'd imagine would be a reason the user would use this.